api, move http api out of rpm package
